See full commit message for details. Revolt.Net/Prevent glitched attachments from failing JSON parsing Revolt.Net/Permissions Revolt.Net/Add various things Taco/Tests with permissions Taco/Use CommandContext for precondition attributes instead of Message DiscordBridge/Fix casing

+++ Revolt.Net/Attachment.cs
@@ -14,7 +14,7 @@ namespace Revolt
     public class AttachmentMetadata
     {
         [JsonProperty("type")] public string Type { get; private set; }
-        [JsonProperty("width")] public int? Width { get; private set; }
-        [JsonProperty("height")] public int? Height { get; private set; }
+        [JsonProperty("width")] public long? Width { get; private set; }
+        [JsonProperty("height")] public long? Height { get; private set; }
     }
 }
